+++ src/Crowdin.Api/Reports/TranslationCostsPostEditingGenerateReportRequest.cs
@@ -76,7 +76,7 @@ namespace Crowdin.Api.Reports
         public class ByTaskSchema : SchemaBase
         {
             [JsonProperty("taskId")]
-            public int? TaskId { get; set; }
+            public long? TaskId { get; set; }
         }
 
         [PublicAPI]
