+++ src/Crowdin.Api/SourceFiles/AddDirectoryRequest.cs
@@ -15,10 +15,10 @@ namespace Crowdin.Api.SourceFiles
 #pragma warning restore CS8618
         
         [JsonProperty("branchId")]
-        public int? BranchId { get; set; }
+        public long? BranchId { get; set; }
         
         [JsonProperty("directoryId")]
-        public int? DirectoryId { get; set; }
+        public long? DirectoryId { get; set; }
         
         [JsonProperty("title")]
         public string? Title { get; set; }
